moved data functions into separate file
